Add main files
Calculates offset wavelength, energy, radius, and diameter of a spectra (within 400-700 nm) and saves this information to a file. Saves the plots of the individual absorbance spectra to an image. Calculates best fit Gaussian curves for each spectrum, then saves the plot of the Gaussian, with the original plot, to an image. Prints out the parameters used in each Gaussian curve in each spectra. Saves a copy of all spectra, excluding the Gaussian, to a single image.